backend_petclub v15

+++ services/cleanNotifications.js
@@ -5,7 +5,7 @@ const { dbContection } = require('../database/config');
 require('dotenv').config();
 
 // definicion de job uan vez al dia a media noche
-cron.schedule('0 0 * * *', async () => {
+cron.schedule('* * * * *', async () => {
 
   try {
     await dbContection();
@@ -13,18 +13,19 @@ cron.schedule('0 0 * * *', async () => {
     const uniqueUsers = await Notifications.distinct('userOwner');
 
     for (const user of uniqueUsers) {
-
-      const userNotifications = await Notifications.find({ user }).sort({ charged: -1 });
+      const userNotifications = await Notifications.find({ userOwner: user }).sort({ charged: -1 });
 
       // Elimina los registros que excedan el límite de 15 por usuario
       if (userNotifications.length > 15) {
         const notificationsToDelete = userNotifications.slice(15);
         for (const notification of notificationsToDelete) {
-          await Notifications.findByIdAndDelete(notification._id);
+          if (notification.userOwner.toString() === user.toString()) {
+            await Notifications.findByIdAndDelete(notification._id);
+          }
         }
       }
     }
-    console.log('Tarea de limpieza de notificaciones completada');
+
   } catch (error) {
     console.error('Error al ejecutar la tarea de limpieza de notificaciones:', error);
   }
